Sprint-0/01-WebApi-Conversion - moving Registration and Login logic into the User area. Also, creating exeption catch handling for when there is no user present in the system.
