+++ AngularFrontEnd/AngularFrontEnd.csproj
@@ -85,11 +85,10 @@
     <Content Include="site\app\example\exampleList.component.html" />
     <Content Include="site\app\example\exampleList.component.js" />
     <Content Include="site\app\example\exampleSingleton.service.js" />
-    <Content Include="site\app\login\login.component.html" />
-    <Content Include="site\app\login\login.component.js" />
-    <Content Include="site\app\register\register.component.html" />
-    <Content Include="site\app\register\register.component.js" />
     <Content Include="site\app\user\login.component.html" />
+    <Content Include="site\app\user\login.component.js" />
+    <Content Include="site\app\user\register.component.html" />
+    <Content Include="site\app\user\register.component.js" />
     <Content Include="site\index.html" />
     <Content Include="site\styles.css" />
     <Content Include="site\systemjs-angular-loader.js" />
@@ -121,8 +120,8 @@
     <Content Include="src\styl\_footer.styl" />
     <Content Include="src\styl\_header.styl" />
     <None Include="packages.config" />
-    <Content Include="site\app\register\register.component.js.map" />
-    <Content Include="site\app\login\login.component.js.map" />
+    <Content Include="site\app\user\register.component.js.map" />
+    <Content Include="site\app\user\login.component.js.map" />
     <None Include="Web.Debug.config">
       <DependentUpon>Web.config</DependentUpon>
     </None>
@@ -148,13 +147,12 @@
     <TypeScriptCompile Include="site\app\example\exampleList.component.ts" />
     <TypeScriptCompile Include="site\app\example\exampleSingleton.service.ts" />
     <TypeScriptCompile Include="site\app\home\home.component.ts" />
-    <TypeScriptCompile Include="site\app\login\login.component.ts" />
-    <TypeScriptCompile Include="site\app\register\register.component.ts" />
+    <TypeScriptCompile Include="site\app\user\login.component.ts" />
+    <TypeScriptCompile Include="site\app\user\register.component.ts" />
     <TypeScriptCompile Include="site\app\shared\preloader.service.ts" />
     <TypeScriptCompile Include="site\app\user\auth-route-guard.service.ts" />
     <TypeScriptCompile Include="site\app\shared\pageNotFound.component.ts" />
     <TypeScriptCompile Include="site\app\user\auth.service.ts" />
-    <TypeScriptCompile Include="site\app\user\login.component.ts" />
     <TypeScriptCompile Include="site\app\user\token-interceptor.service.ts" />
     <TypeScriptCompile Include="site\app\user\user.module.ts" />
     <TypeScriptCompile Include="site\app\user\user.ts" />
