+++ AngularFrontEnd/site/app/app.module.ts
@@ -3,8 +3,8 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { AppComponent } from './app.component';
 import { ExampleListComponent } from './example/exampleList.component';
-import { RegisterComponent } from './register/register.component';
-import { LoginComponent } from './login/login.component';
+import { RegisterComponent } from './user/register.component';
+import { LoginComponent } from './user/login.component';
 import { ExampleComponent } from './example/example.component';
 import { ExampleCountComponent } from './example/exampleCount.component';
 import { HttpModule } from '@angular/http';
