--- AngularFrontEnd/site/app/login/login.component.html
@@ -1,29 +1,0 @@
-<div class="py-5">
-    <div class="row">
-        <div class="col-md-6 mx-auto">
-            <span class="anchor" id="formLogin"></span>
-            <!-- form card login -->
-            <div class="card rounded-0">
-                <div class="card-header">
-                    <h3 class="mb-0">Login</h3>
-                </div>
-                <div class="card-body">
-                    <form class="form">
-                        <div class="form-group">
-                            <label for="uname1">Username</label>
-                            <input type="text" class="form-control rounded-0" [(ngModel)]="loginUserInfo.username" name="username" placeholder="place holder text" required>
-                        </div>
-                        <div class="form-group">
-                            <label>Password</label>
-                            <input type="password" class="form-control rounded-0" [(ngModel)]="loginUserInfo.password" name="password" required>
-                        </div>
-                        <button type="button" (click)="loginUser()" class="btn btn-success float-right">Login</button>
-                    </form>
-                </div>
-                <!--/card-block-->
-            </div>
-            <!-- /form card login -->
-        </div>
-
-    </div>
-</div>
