+++ AngularFrontEnd/site/app/user/login.component.html
@@ -1,87 +1,29 @@
-<div class="panel panel-primary">
-    <div class="panel-heading">
-        {{pageTitle}}
+<div class="py-5">
+    <div class="row">
+        <div class="col-md-6 mx-auto">
+            <span class="anchor" id="formLogin"></span>
+            <!-- form card login -->
+            <div class="card rounded-0">
+                <div class="card-header">
+                    <h3 class="mb-0">Login</h3>
+                </div>
+                <div class="card-body">
+                    <form class="form">
+                        <div class="form-group">
+                            <label for="uname1">Username</label>
+                            <input type="text" class="form-control rounded-0" [(ngModel)]="loginUserInfo.username" name="username" placeholder="place holder text" required>
+                        </div>
+                        <div class="form-group">
+                            <label>Password</label>
+                            <input type="password" class="form-control rounded-0" [(ngModel)]="loginUserInfo.password" name="password" required>
+                        </div>
+                        <button type="button" (click)="loginUser()" class="btn btn-success float-right">Login</button>
+                    </form>
+                </div>
+                <!--/card-block-->
+            </div>
+            <!-- /form card login -->
+        </div>
+
     </div>
-
-    <div class="panel-body">
-        <form class="form-horizontal"
-              novalidate
-              (ngSubmit)="login(loginForm)"
-              #loginForm="ngForm"
-              autocomplete="off" >
-            <fieldset>
-                <div class="form-group" 
-                     [ngClass]="{'has-error': (userNameVar.touched || 
-                                               userNameVar.dirty) && 
-                                               !userNameVar.valid }">
-                    <label class="col-md-2 control-label" 
-                           for="userNameId">User Name</label>
-
-                    <div class="col-md-8">
-                        <input class="form-control" 
-                                id="userNameId" 
-                                type="text" 
-                                placeholder="User Name (required)"
-                                required
-                                (ngModel) = "userName"
-                                name="userName"
-                                #userNameVar="ngModel" />
-                        <span class="help-block" *ngIf="(userNameVar.touched ||
-                                                         userNameVar.dirty) &&
-                                                         userNameVar.errors">
-                            <span *ngIf="userNameVar.errors.required">
-                                User name is required.
-                            </span>
-                        </span>
-                    </div>
-                </div>
-                
-                <div class="form-group" 
-                     [ngClass]="{'has-error': (passwordVar.touched || 
-                                               passwordVar.dirty) && 
-                                               !passwordVar.valid }">
-                    <label class="col-md-2 control-label" for="passwordId">Password</label>
-
-                    <div class="col-md-8">
-                        <input class="form-control" 
-                                id="passwordId" 
-                                type="password" 
-                                placeholder="Password (required)"
-                                required
-                                (ngModel) = "password"
-                                name="password"
-                                #passwordVar="ngModel" />
-                        <span class="help-block" *ngIf="(passwordVar.touched ||
-                                                         passwordVar.dirty) &&
-                                                         passwordVar.errors">
-                            <span *ngIf="passwordVar.errors.required">
-                                Password is required.
-                            </span>
-                        </span>
-                    </div>
-                </div>
-                
-                <div class="form-group">
-                    <div class="col-md-4 col-md-offset-2">
-                        <span>
-                            <button class="btn btn-primary"
-                                    type="submit"
-                                    style="width:80px;margin-right:10px"
-                                    [disabled]="!loginForm.valid">
-                                Log In
-                            </button>
-                        </span>
-                        <span>
-                            <a class="btn btn-default"
-                               [routerLink]="['/welcome']">
-                                Cancel
-                            </a>
-                        </span>
-                     </div>
-                </div>
-            </fieldset>
-        </form>
-        <div class="has-error" *ngIf="errorMessage">{{errorMessage}}</div>
-    </div>
-
 </div>
