+++ AngularFrontEnd/site/app/user/login.component.ts
@@ -1,30 +1,32 @@
-import { Component } from '@angular/core';
-import { NgForm } from '@angular/forms';
-import { Router } from '@angular/router';
-import { AuthService } from './auth.service';
+import { Component, OnInit } from '@angular/core';
+import { AuthService } from '../user/auth.service';
+import { Router } from '@angular/router'
+import { IUser } from '../user/user';
 
 @Component({
+    selector: 'app-login',
     templateUrl: './login.component.html'
 })
-export class LoginComponent {
-    errorMessage: string;
-    pageTitle = 'Please Log In';
+export class LoginComponent implements OnInit {
 
-    constructor(private authService: AuthService, private router: Router) { }
+    loginUserInfo: IUser = { username: '', password: '' } as IUser;
 
-    login(loginForm: NgForm) {
-        if (loginForm && loginForm.valid) {
-            let userName = loginForm.form.value.userName;
-            let password = loginForm.form.value.password;
-            //this.authService.login(userName, password);
+    constructor(private _auth: AuthService, private _router: Router) { }
 
-            if (this.authService.redirectUrl) {
-                this.router.navigateByUrl(this.authService.redirectUrl);
-            } else {
-                this.router.navigate(['/home']);
-            }
-        } else {
-            this.errorMessage = 'Please enter a user name and password.';
-        };
+    ngOnInit() { }
+
+    loginUser() {
+        this._auth.loginUser(this.loginUserInfo)
+            .subscribe(
+            res => {
+                localStorage.setItem('token', res.token)
+                this._router.navigate(['/home'])
+            },
+            err => {
+                localStorage.removeItem('token')//probably wont have a token anyway, but whatever.
+
+                //TODO - do something here to let the user Know the Username/Password is INVALID for some reason
+            });
     }
+
 }
